removed user signUp data to contxt.js file
I moved the data from the sigUp.jsx file and moved it to a userAccountContext.js file so that I can make the signup file its  own component to be also used for sign in.
