+++ src/App.js
@@ -1,9 +1,10 @@
+import { UserAccountProvider } from "./context/userAccountContext"
 import SignUp from "./components/SignUp"
 function App() {
   return (
-    <>
+    <UserAccountProvider>
       <SignUp />
-    </>
+    </UserAccountProvider>
   )
 }
