+++ src/components/SignUp.jsx
@@ -1,23 +1,19 @@
-import { useState } from "react"
+import { useState, useContext } from "react"
+import UserAccountContext from "../context/userAccountContext"
 function SignUp() {
-  const [name, setName] = useState("")
-  const [email, setEmail] = useState("")
-  const [password, setPassword] = useState("")
-  const [passwordConfirm, setPasswordConfirm] = useState("")
-  const [error, setError] = useState("")
+  const {
+    name,
+    email,
+    password,
+    passwordConfirm,
+    error,
+    handleName,
+    handleEmail,
+    handlePassword,
+    handlePasswordConfirm,
+    setError,
+  } = useContext(UserAccountContext)
 
-  const handleEmail = (e) => {
-    setEmail(e.target.value)
-  }
-  const handlePassword = (e) => {
-    setPassword(e.target.value)
-  }
-  const handlePasswordConfirm = (e) => {
-    setPasswordConfirm(e.target.value)
-  }
-  const handleName = (e) => {
-    setName(e.target.value)
-  }
   const handleSubmit = (e) => {
     e.preventDefault()
     const data = { name, email, password, passwordConfirm }
